refactor: 프로젝트명 변경 사항을 각 클래스에 적용
